Abstractions to compile systemlib for Linux and Nuttx
Modified uint32_t casts of pointers to unsigned long for portability.
It otherwise breaks on x86_64.

Added _PX4_IOC to handle the conflice between _IOC on Linux and NuttX.

Removed use of px4::ok() because it cannot be used in a thread based
implementation. Changed to use px4::AppMgr which uses ok() on ROS.

Removed up_cxxinitialize.c from Linux build.

+++ src/modules/systemlib/perf_counter.c
@@ -386,7 +386,7 @@ perf_print_counter_fd(int fd, perf_counter_t handle)
 	case PC_COUNT:
 		dprintf(fd, "%s: %llu events\n",
 		       handle->name,
-		       ((struct perf_ctr_count *)handle)->event_count);
+		       (unsigned long long)((struct perf_ctr_count *)handle)->event_count);
 		break;
 
 	case PC_ELAPSED: {
@@ -395,12 +395,12 @@ perf_print_counter_fd(int fd, perf_counter_t handle)
 
 		dprintf(fd, "%s: %llu events, %llu overruns, %lluus elapsed, %lluus avg, min %lluus max %lluus %5.3fus rms\n",
 			handle->name,
-			pce->event_count,
-			pce->event_overruns,
-			pce->time_total,
-			pce->time_total / pce->event_count,
-			pce->time_least,
-			pce->time_most,
+			(unsigned long long)pce->event_count,
+			(unsigned long long)pce->event_overruns,
+			(unsigned long long)pce->time_total,
+			(unsigned long long)pce->time_total / pce->event_count,
+			(unsigned long long)pce->time_least,
+			(unsigned long long)pce->time_most,
 			(double)(1e6f * rms));
 		break;
 	}
@@ -411,10 +411,10 @@ perf_print_counter_fd(int fd, perf_counter_t handle)
 
 		dprintf(fd, "%s: %llu events, %lluus avg, min %lluus max %lluus %5.3fus rms\n",
 			handle->name,
-			pci->event_count,
-			(pci->time_last - pci->time_first) / pci->event_count,
-			pci->time_least,
-			pci->time_most,
+			(unsigned long long)pci->event_count,
+			(unsigned long long)(pci->time_last - pci->time_first) / pci->event_count,
+			(unsigned long long)pci->time_least,
+			(unsigned long long)pci->time_most,
 			(double)(1e6f * rms));
 		break;
 	}
